services/doctor_service/requirements.txt
Django>=3.2,<4.3
djangorestframework>=3.12,<3.15
djangorestframework-simplejwt>=5.0,<5.4
requests
django-cors-headers>=3.0,<4.0